refactor: Refactor playground and signal modules
Refactored the playground and signal modules to use the useGlobalContext hook instead of the useRouter hook.
