+++ playground/main.tsx
@@ -1,7 +1,7 @@
 import "@radix-ui/themes/styles.css";
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { useRouter } from "./Router";
+import { useGlobalContext } from "./useGlobalContext";
 import { Box, Container, Theme } from "@radix-ui/themes";
 import { App } from "./App";
 
@@ -17,9 +17,9 @@ ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
     >
       <Box p="8" style={{ backgroundColor: "var(--gray-1)" }} height="100%">
         <Container>
-          <useRouter.Provider>
+          <useGlobalContext.Provider>
             <App />
-          </useRouter.Provider>
+          </useGlobalContext.Provider>
         </Container>
       </Box>
     </Theme>
